Update and refactor
