+++ recipes/ESC50/classification/extra_requirements.txt
@@ -1,5 +1,3 @@
 matplotlib
-pandas
 scikit-learn
 torchvision
-transformers>=4.29.1
