+++ recipes/ESC50/classification/hparams/cnn14.yaml
@@ -1,11 +1,10 @@
 # #################################
 # Basic training parameters for sound classification using the ESC50 dataset.
-# This recipe uses a FocalNet backbone for classification.
+# This recipe uses the ecapa-tdnn backbone for classification.
 #
 # Authors:
 #  * Cem Subakan 2022, 2023
 #  * Francesco Paissan 2022, 2023
-#  * Luca Della Libera 2024
 #  (based on the SpeechBrain UrbanSound8k recipe)
 # #################################
 
@@ -14,14 +13,14 @@ seed: 1234
 __set_seed: !!python/object/apply:torch.manual_seed [!ref <seed>]
 
 # Set up folders for reading from and writing to
-# Dataset must already exist at `audio_data_folder`
-data_folder: !PLACEHOLDER # e.g., /localscratch/ESC-50-master
+data_folder: !PLACEHOLDER  # e.g., /localscratch/ESC-50-master
 audio_data_folder: !ref <data_folder>/audio
 
-experiment_name: focalnet-base-esc50
+experiment_name: cnn14-esc50
 output_folder: !ref ./results/<experiment_name>/<seed>
 save_folder: !ref <output_folder>/save
 train_log: !ref <output_folder>/train_log.txt
+
 
 # Tensorboard logs
 use_tensorboard: False
@@ -42,13 +41,23 @@ skip_manifest_creation: False
 ckpt_interval_minutes: 15 # save checkpoint every N min
 
 # Training parameters
-number_of_epochs: 100
-batch_size: 16
+number_of_epochs: 200
+batch_size: 32
 lr: 0.0002
 base_lr: 0.00000001
 max_lr: !ref <lr>
 step_size: 65000
-sample_rate: 16000
+sample_rate: 44100
+
+device: "cpu"
+
+# Feature parameters
+n_mels: 80
+left_frames: 0
+right_frames: 0
+deltas: False
+
+use_melspectra: True
 
 # Number of classes
 out_n_neurons: 50
@@ -60,55 +69,49 @@ dataloader_options:
     shuffle: !ref <shuffle>
     num_workers: 0
 
-# Augmentation
-drop_freq: !new:speechbrain.augment.time_domain.DropFreq
-    drop_freq_low: 0  # Min frequency band dropout probability
-    drop_freq_high: 1  # Max frequency band dropout probability
-    drop_freq_count_low: 1  # Min number of frequency bands to drop
-    drop_freq_count_high: 3  # Max number of frequency bands to drop
-    drop_freq_width: 0.05  # Width of frequency bands to drop
+# Functions
+compute_features: !new:speechbrain.lobes.features.Fbank
+    n_mels: !ref <n_mels>
+    left_frames: !ref <left_frames>
+    right_frames: !ref <right_frames>
+    deltas: !ref <deltas>
+    sample_rate: !ref <sample_rate>
+    n_fft: 1024
+    win_length: 20
+    hop_length: 10
 
-drop_chunk: !new:speechbrain.augment.time_domain.DropChunk
-    drop_length_low: 1  # Min number of audio chunks to drop
-    drop_length_high: 5  # Max number of audio chunks to drop
-    drop_count_low: 1000  # Min length of audio chunks to drop
-    drop_count_high: 2000  # Max length of audio chunks to drop
-
-augmentation: !new:speechbrain.augment.augmenter.Augmenter
-    parallel_augment: False
-    concat_original: False
-    repeat_augment: 1
-    shuffle_augmentations: False
-    min_augmentations: 2
-    max_augmentations: 2
-    augment_prob: 0.75
-    augmentations: [!ref <drop_freq>, !ref <drop_chunk>]
-
-# Model
-embedding_model: !apply:transformers.FocalNetBackbone.from_pretrained [microsoft/focalnet-base]
+embedding_model: !new:speechbrain.lobes.models.Cnn14.Cnn14
+    mel_bins: !ref <n_mels>
+    emb_dim: 2048
 
 classifier: !new:speechbrain.lobes.models.ECAPA_TDNN.Classifier
-    input_size: 1024
+    input_size: 2048
     out_neurons: !ref <out_n_neurons>
     lin_blocks: 1
 
 epoch_counter: !new:speechbrain.utils.epoch_loop.EpochCounter
     limit: !ref <number_of_epochs>
 
-# Pre-processing
+# pre-processing
 n_fft: 1024
 spec_mag_power: 0.5
 hop_length: 11.6099
 win_length: 23.2199
-
 compute_stft: !new:speechbrain.processing.features.STFT
     n_fft: !ref <n_fft>
     hop_length: !ref <hop_length>
     win_length: !ref <win_length>
     sample_rate: !ref <sample_rate>
 
+compute_fbank: !new:speechbrain.processing.features.Filterbank
+    n_mels: 80
+    n_fft: !ref <n_fft>
+    sample_rate: !ref <sample_rate>
+
 modules:
     compute_stft: !ref <compute_stft>
+    compute_fbank: !ref <compute_fbank>
+    compute_features: !ref <compute_features>
     embedding_model: !ref <embedding_model>
     classifier: !ref <classifier>
 
@@ -140,3 +143,13 @@ checkpointer: !new:speechbrain.utils.checkpoints.Checkpointer
         embedding_model: !ref <embedding_model>
         classifier: !ref <classifier>
         counter: !ref <epoch_counter>
+
+use_pretrained: True
+# If you do not want to use the pretrained encoder you can simply delete pretrained_encoder field.
+embedding_model_path: speechbrain/cnn14-esc50/embedding_model.ckpt
+pretrained_encoder: !new:speechbrain.utils.parameter_transfer.Pretrainer
+    collect_in: !ref <save_folder>
+    loadables:
+        embedding_model: !ref <embedding_model>
+    paths:
+        embedding_model: !ref <embedding_model_path>
