+++ recipes/ESC50/classification/train.py
@@ -1,9 +1,9 @@
 #!/usr/bin/python3
 
-"""Recipe to train a ViT classifier on ESC50 data.
+"""Recipe to train a classifier on ESC50 data.
 
 To run this recipe, use the following command:
-> python train_vit_classifier.py hparams/vit_classifier.yaml --data_folder yourpath/ESC-50-master
+> python train.py hparams/<config>.yaml --data_folder yourpath/ESC-50-master
 
 Authors
     * Cem Subakan 2022, 2023
@@ -34,14 +34,12 @@ class ESC50Brain(sb.core.Brain):
     """Class for classifier training."""
 
     def compute_forward(self, batch, stage):
-        """Computation pipeline based on an encoder + sound classifier.
-        Data augmentation is applied to the input sound.
-        """
+        """Computation pipeline based on an encoder + sound classifier."""
         batch = batch.to(self.device)
         wavs, lens = batch.sig
 
         # Augment if specified
-        if stage == sb.Stage.TRAIN:
+        if hasattr(self.hparams, "augmentation") and stage == sb.Stage.TRAIN:
             wavs, lens = self.hparams.augmentation(wavs, lens)
 
         # Extract features
@@ -49,27 +47,48 @@ class ESC50Brain(sb.core.Brain):
         X_stft_power = sb.processing.features.spectral_magnitude(
             X_stft, power=self.hparams.spec_mag_power
         )
-        net_input = torch.log1p(X_stft_power)
+        if (
+            hasattr(self.hparams, "use_melspectra")
+            and self.hparams.use_melspectra
+        ):
+            net_input = self.modules.compute_fbank(X_stft_power)
+        else:
+            net_input = torch.log1p(X_stft_power)
 
         # Embeddings + sound classifier
-        image_size = self.modules.embedding_model.config.image_size
-        net_input = torchvision.transforms.functional.resize(
-            net_input, (image_size, image_size)
-        )
-        net_input = net_input[:, None, ...].expand(
-            -1, 3, -1, -1
-        )  # Expand to have 3 channels
-        embeddings = self.modules.embedding_model(
-            net_input
-        ).last_hidden_state.movedim(-1, -2)
-        embeddings = embeddings.mean(dim=-1)
+        if hasattr(self.modules.embedding_model, "config"):
+            # Hugging Face model
+            config = self.modules.embedding_model.config
+            # Resize to match expected resolution
+            net_input = torchvision.transforms.functional.resize(
+                net_input, (config.image_size, config.image_size)
+            )
+            # Expand to have 3 channels
+            net_input = net_input[:, None, ...].expand(-1, 3, -1, -1)
+            if config.model_type == "focalnet":
+                embeddings = self.modules.embedding_model(
+                    net_input
+                ).feature_maps[-1]
+                embeddings = embeddings.mean(dim=(-1, -2))
+            elif config.model_type == "vit":
+                embeddings = self.modules.embedding_model(
+                    net_input
+                ).last_hidden_state.movedim(-1, -2)
+                embeddings = embeddings.mean(dim=-1)
+            else:
+                raise NotImplementedError
+        else:
+            # SpeechBrain model
+            embeddings = self.modules.embedding_model(net_input)
+            if embeddings.ndim == 4:
+                embeddings = embeddings.mean((-1, -2))
+
         outputs = self.modules.classifier(embeddings)
 
         return outputs, lens
 
     def compute_objectives(self, predictions, batch, stage):
-        """Computes the loss using class-id as label.
-        """
+        """Computes the loss using class-id as label."""
         predictions, lens = predictions
         uttid = batch.id
         classid, _ = batch.class_string_encoded
@@ -80,7 +99,7 @@ class ESC50Brain(sb.core.Brain):
             if hasattr(self.hparams.lr_annealing, "on_batch_end"):
                 self.hparams.lr_annealing.on_batch_end(self.optimizer)
 
-        # Append this batch of losses to the loss metric for easy
+        # Append this batch of losses to the loss metric
         self.loss_metric.append(
             uttid, predictions, classid, lens, reduction="batch"
         )
@@ -105,7 +124,7 @@ class ESC50Brain(sb.core.Brain):
             )
             self.test_confusion_matrix += confusion_matix
 
-        # Compute Accuracy using MetricStats
+        # Compute accuracy using MetricStats
         self.acc_metric.append(
             uttid, predict=predictions, target=classid, lengths=lens
         )
@@ -131,10 +150,10 @@ class ESC50Brain(sb.core.Brain):
             metric=sb.nnet.losses.nll_loss
         )
 
-        # Compute Accuracy using MetricStats
+        # Compute accuracy using MetricStats
         # Define function taking (prediction, target, length) for eval
         def accuracy_value(predict, target, lengths):
-            """Computes Accuracy"""
+            """Computes accuracy."""
             nbr_correct, nbr_total = sb.utils.Accuracy.Accuracy(
                 predict, target, lengths
             )
@@ -181,16 +200,14 @@ class ESC50Brain(sb.core.Brain):
                 "loss": self.train_loss,
                 "acc": self.acc_metric.summarize("average"),
             }
-        # Summarize Valid statistics from the stage for record-keeping.
+        # Summarize Valid statistics from the stage for record-keeping
         elif stage == sb.Stage.VALID:
             valid_stats = {
                 "loss": stage_loss,
-                "acc": self.acc_metric.summarize(
-                    "average"
-                ),  # "acc": self.valid_acc_metric.summarize(),
+                "acc": self.acc_metric.summarize("average"),
                 "error": self.error_metrics.summarize("average"),
             }
-        # Summarize Test statistics from the stage for record-keeping.
+        # Summarize Test statistics from the stage for record-keeping
         else:
             test_stats = {
                 "loss": stage_loss,
@@ -221,7 +238,7 @@ class ESC50Brain(sb.core.Brain):
                     "Validation Confusion Matrix", cm_fig, epoch
                 )
 
-            # The train_logger writes a summary to stdout and to the logfile.
+            # The train_logger writes a summary to stdout and to the log file
             self.hparams.train_logger.log_stats(
                 stats_meta={"epoch": epoch, "lr": old_lr},
                 train_stats=self.train_stats,
@@ -232,7 +249,7 @@ class ESC50Brain(sb.core.Brain):
                 meta=valid_stats, min_keys=["error"]
             )
 
-        # We also write statistics about test data to stdout and to the logfile.
+        # We also write statistics about test data to stdout and to the log file
         if stage == sb.Stage.TEST:
             # Per class accuracy from Test confusion matrix
             per_class_acc_arr = np.diag(self.test_confusion_matrix) / np.sum(
@@ -298,6 +315,7 @@ def dataio_prep(hparams):
     @sb.utils.data_pipeline.takes("class_string")
     @sb.utils.data_pipeline.provides("class_string", "class_string_encoded")
     def label_pipeline(class_string):
+        """The label pipeline."""
         yield class_string
         class_string_encoded = label_encoder.encode_label_torch(class_string)
         yield class_string_encoded
@@ -332,6 +350,9 @@ def dataio_prep(hparams):
 
 
 if __name__ == "__main__":
+    # This flag enables the built-in cuDNN auto-tuner
+    # torch.backends.cudnn.benchmark = True
+
     # CLI:
     hparams_file, run_opts, overrides = sb.parse_arguments(sys.argv[1:])
