+++ recipes/ESC50/interpret/hparams/amt_vit.yaml
@@ -1,5 +1,6 @@
 # #################################
-# Recipe to interpret a ViT audio classifier by-design based on its attention maps.
+# Recipe to interpret a ViT audio classifier by-design via activation map thresholding (AMT)
+# applied to its attention maps.
 #
 # Author:
 #  * Cem Subakan 2022, 2023
@@ -13,8 +14,7 @@ seed: 1234
 __set_seed: !!python/object/apply:torch.manual_seed [!ref <seed>]
 
 # Set up folders for reading from and writing to
-# Dataset must already exist at `audio_data_folder`
-data_folder: !PLACEHOLDER # e.g., /localscratch/UrbanSound8K
+data_folder: !PLACEHOLDER  # e.g., /localscratch/ESC-50-master
 audio_data_folder: !ref <data_folder>/audio
 
 experiment_name: vit-base
@@ -85,8 +85,8 @@ classifier: !new:speechbrain.lobes.models.ECAPA_TDNN.Classifier
     out_neurons: !ref <out_n_neurons>
     lin_blocks: 1
 
-embedding_model_path: null
-classifier_model_path: null
+embedding_model_path: !PLACEHOLDER
+classifier_model_path: !PLACEHOLDER
 
 pretrained_esc50: !new:speechbrain.utils.parameter_transfer.Pretrainer
     collect_in: !ref <save_folder>
