+++ recipes/ESC50/interpret/hparams/piq_focalnet.yaml
@@ -13,8 +13,7 @@ seed: 1234
 __set_seed: !!python/object/apply:torch.manual_seed [!ref <seed>]
 
 # Set up folders for reading from and writing to
-# Dataset must already exist at `audio_data_folder`
-data_folder: !PLACEHOLDER # e.g., /localscratch/UrbanSound8K
+data_folder: !PLACEHOLDER  # e.g., /localscratch/ESC-50-master
 audio_data_folder: !ref <data_folder>/audio
 
 experiment_name: piq_focalnet-base
@@ -129,7 +128,7 @@ compute_istft: !new:speechbrain.processing.features.ISTFT
     hop_length: !ref <hop_length>
     win_length: !ref <win_length>
 
-psi_model: !new:train_piq_focalnet.VectorQuantizedPSIFocalNet
+psi_model: !new:train_piq.VectorQuantizedPSIFocalNet
     dim: 1024
     K: !ref <K>
     shared_keys: 0
