+++ recipes/ESC50/interpret/interpret_amt.py
@@ -1,6 +1,9 @@
 #!/usr/bin/python3
 
-"""Recipe to interpret a FocalNet audio classifier by-design based on its modulation maps.
+"""Recipe to interpret an audio classifier by-design via activation maps thresholding (AMT).
+
+To run this recipe, use the following command:
+> python intepret_amt.py hparams/<amt-config>.yaml --data_folder /yourpath/ESC-50-master
 
 Authors
     * Cem Subakan 2022, 2023
@@ -10,6 +13,7 @@ Authors
 
 import os
 import sys
+
 import matplotlib.pyplot as plt
 import speechbrain as sb
 import torch
@@ -23,7 +27,7 @@ from torch.nn import functional as F
 
 
 class InterpreterESC50Brain(sb.core.Brain):
-    """Class for sound class embedding training" """
+    """Class for interpreter training."""
 
     def invert_stft_with_phase(self, X_int, X_stft_phase):
         """Inverts STFT spectra given phase."""
@@ -53,36 +57,68 @@ class InterpreterESC50Brain(sb.core.Brain):
 
         return X_stft_logpower, X_stft, X_stft_power
 
+    @torch.no_grad()
     def classifier_forward(self, X_stft_logpower):
-        """The forward pass for the classifier"""
-        image_size = self.hparams.embedding_model.config.image_size
+        """The forward pass for the classifier."""
+        config = self.modules.embedding_model.config
+        # Resize to match expected resolution
         net_input = torchvision.transforms.functional.resize(
-            X_stft_logpower, (image_size, image_size)
+            X_stft_logpower, (config.image_size, config.image_size)
         )
-        net_input = net_input[:, None, ...].expand(
-            -1, 3, -1, -1
-        )  # Expand to have 3 channels
-        hcat = self.hparams.embedding_model(net_input).feature_maps[-1]
-        embeddings = hcat.mean(dim=(-1, -2))
+        # Expand to have 3 channels
+        net_input = net_input[:, None, ...].expand(-1, 3, -1, -1)
+        if config.model_type == "focalnet":
+            hcat = self.hparams.embedding_model(net_input).feature_maps[-1]
+            embeddings = hcat.mean(dim=(-1, -2))
+            modulators = [
+                encoder_stage.layers[-1].modulation.modulator
+                for encoder_stage in self.hparams.embedding_model.focalnet.encoder.stages
+            ]
+            modulators = [x.norm(dim=-3, p=2, keepdim=True) for x in modulators]
+            # Upsample spatial dimensions
+            modulators = [
+                torchvision.transforms.functional.resize(
+                    x, X_stft_logpower.shape[-2:]
+                )
+                for x in modulators
+            ]
+            xhat = modulators[-1]
+        elif config.model_type == "vit":
+            model_output = self.hparams.embedding_model(
+                net_input, output_attentions=True
+            )
+            hcat = model_output.last_hidden_state.movedim(-1, -2)
+            embeddings = hcat.mean(dim=-1)
+
+            # Take the representations from CLS token
+            num_heads = self.hparams.embedding_model.config.num_attention_heads
+            attentions = [x[:, :, 0, 1:] for x in model_output.attentions]
+
+            # Reshape the attention scores to resemble mini patches
+            num_patches = (
+                self.hparams.embedding_model.config.image_size
+                // self.hparams.embedding_model.config.patch_size
+            )
+
+            attentions = [
+                x.reshape(-1, num_heads, num_patches, num_patches)
+                for x in attentions
+            ]
+            attentions = [x.mean(dim=-3, keepdim=True) for x in attentions]
+            # Upsample spatial dimensions
+            attentions = [
+                torchvision.transforms.functional.resize(
+                    x, X_stft_logpower.shape[-2:]
+                )
+                for x in attentions
+            ]
+            xhat = attentions[-1]
+        else:
+            raise NotImplementedError
+
         predictions = self.hparams.classifier(embeddings).squeeze(1)
         class_pred = predictions.argmax(1)
 
-        modulators = [
-            encoder_stage.layers[-1].modulation.modulator
-            for encoder_stage in self.hparams.embedding_model.focalnet.encoder.stages
-        ]
-
-        modulators = [x.norm(dim=-3, p=2, keepdim=True) for x in modulators]
-
-        modulators = [
-            torchvision.transforms.functional.resize(
-                x, X_stft_logpower.shape[-2:]
-            )
-            for x in modulators
-        ]
-
-        xhat = modulators[-1]
-        # xhat = (xhat - xhat.mean(dim=(-1, -2), keepdim=True))
         threshold = xhat.reshape(len(xhat), -1).quantile(
             self.hparams.quantile, dim=-1
         )[:, None, None, None]
@@ -92,7 +128,7 @@ class InterpreterESC50Brain(sb.core.Brain):
         return xhat, predictions, class_pred
 
     def interpret_computation_steps(self, wavs, print_probability=False):
-        """Computation steps to get the interpretation spectrogram"""
+        """Computation steps to get the interpretation spectrogram."""
         X_stft_logpower, X_stft, X_stft_power = self.preprocess(wavs)
         X_stft_phase = spectral_phase(X_stft)
 
@@ -118,7 +154,7 @@ class InterpreterESC50Brain(sb.core.Brain):
     def interpret_sample(self, wavs, batch=None):
         """Get the interpratation for a given wav file."""
 
-        # get the interpretation spectrogram, phase, and the predicted class
+        # Get the interpretation spectrogram, phase, and the predicted class
         X_int, X_stft_phase, pred_cl, _, _ = self.interpret_computation_steps(
             wavs
         )
@@ -126,7 +162,7 @@ class InterpreterESC50Brain(sb.core.Brain):
         if not (batch is None):
             x_int_sb = self.invert_stft_with_phase(X_int, X_stft_phase)
 
-            # save reconstructed and original spectrograms
+            # Save reconstructed and original spectrograms
             os.makedirs(
                 os.path.join(
                     self.hparams.output_folder, "audios_from_interpretation",
@@ -164,7 +200,7 @@ class InterpreterESC50Brain(sb.core.Brain):
         return X_int
 
     def overlap_test(self, batch):
-        """Interpration test with overlapped audio"""
+        """Interpration test with overlapped audio."""
         wavs, _ = batch.sig
         wavs = wavs.to(self.device)
 
@@ -176,8 +212,11 @@ class InterpreterESC50Brain(sb.core.Brain):
         s2 = wavs[1]
         s2 = s2 / s2.max()
 
-        # create the mixture with s2 being the noise (lower gain)
-        if self.hparams.concat_sources:
+        # Create the mixture with s2 being the noise (lower gain)
+        if (
+            hasattr(self.hparams, "concat_sources")
+            and self.hparams.concat_sources
+        ):
             length = min(len(s1), len(s2))
             mid = length // 2
             s1[:mid] = 0.0
@@ -187,7 +226,7 @@ class InterpreterESC50Brain(sb.core.Brain):
             mix = (s1 * 0.8 + (s2 * 0.2)).unsqueeze(0)
         mix = mix / mix.max()
 
-        # get the interpretation spectrogram, phase, and the predicted class
+        # Get the interpretation spectrogram, phase, and the predicted class
         (
             X_int,
             X_stft_phase,
@@ -203,7 +242,7 @@ class InterpreterESC50Brain(sb.core.Brain):
         temp = torch.expm1(X_int).unsqueeze(0).unsqueeze(-1)
         x_int_sb = self.invert_stft_with_phase(temp, X_stft_phase)
 
-        # save reconstructed and original spectrograms
+        # Save reconstructed and original spectrograms
         current_class_ind = batch.class_string_encoded.data[0].item()
         current_class_name = self.hparams.label_encoder.ind2lab[
             current_class_ind
@@ -293,7 +332,7 @@ class InterpreterESC50Brain(sb.core.Brain):
         plt.close()
 
     def debug_files(self, X_stft, xhat, X_stft_logpower, batch, wavs):
-        """The helper function to create debugging images"""
+        """The helper function to create debugging images."""
         X_stft_phase = spectral_phase(X_stft)
         temp = xhat[0].transpose(0, 1).unsqueeze(0).unsqueeze(-1)
         Xspec_est = torch.expm1(temp.permute(0, 2, 1, 3))
@@ -329,7 +368,7 @@ class InterpreterESC50Brain(sb.core.Brain):
         if self.hparams.use_mask_output:
             mask = xhat[0]
         else:
-            mask = xhat[0] > th  # (xhat[0] / xhat[0] + 1e-10)
+            mask = xhat[0] > th
         X_masked = mask * X_stft_logpower[0, :Tmax, :]
         plt.imshow(X_masked.permute(1, 0).data.cpu(), origin="lower")
         plt.colorbar()
@@ -341,7 +380,7 @@ class InterpreterESC50Brain(sb.core.Brain):
         plt.title("mask")
 
         out_folder = os.path.join(
-            self.hparams.output_folder, "reconstructions/" f"{batch.id[0]}",
+            self.hparams.output_folder, "reconstructions", f"{batch.id[0]}",
         )
         os.makedirs(
             out_folder, exist_ok=True,
@@ -371,10 +410,7 @@ class InterpreterESC50Brain(sb.core.Brain):
         )
 
     def compute_forward(self, batch, stage):
-        """Computation pipeline based on an encoder + sound classifier.
-        Data augmentation and environmental corruption are applied to the
-        input sound.
-        """
+        """Computation pipeline based on an encoder + sound classifier."""
         batch = batch.to(self.device)
         wavs, lens = batch.sig
 
@@ -390,7 +426,7 @@ class InterpreterESC50Brain(sb.core.Brain):
         else:
             xhat = F.softplus(xhat)
 
-        # save some samples
+        # Save some samples
         if self.hparams.save_interpretations:
             wavs = wavs[0].unsqueeze(0)
             self.interpret_sample(wavs, batch)
@@ -400,7 +436,7 @@ class InterpreterESC50Brain(sb.core.Brain):
         return predictions, xhat
 
     def compute_objectives(self, pred, batch, stage):
-        """Helper function to compute the objectives"""
+        """Helper function to compute the objectives."""
         predictions, xhat = pred
 
         batch = batch.to(self.device)
@@ -430,11 +466,11 @@ class InterpreterESC50Brain(sb.core.Brain):
         return torch.as_tensor([0.0], device=self.device)
 
     def on_stage_start(self, stage, epoch=None):
-        """Steps taken before stage start"""
+        """Steps taken before stage start."""
 
         @torch.no_grad()
         def accuracy_value(predict, target, length):
-            """Computes Accuracy"""
+            """Computes accuracy."""
             nbr_correct, nbr_total = sb.utils.Accuracy.Accuracy(
                 predict.unsqueeze(1), target, length
             )
@@ -443,7 +479,7 @@ class InterpreterESC50Brain(sb.core.Brain):
 
         @torch.no_grad()
         def compute_fidelity(theta_out, predictions):
-            """Computes top-`k` fidelity of interpreter."""
+            """Computes top-k fidelity of interpreter."""
             predictions = F.softmax(predictions, dim=1)
             theta_out = F.softmax(theta_out, dim=1)
 
@@ -457,7 +493,7 @@ class InterpreterESC50Brain(sb.core.Brain):
 
         @torch.no_grad()
         def compute_faithfulness(wavs, predictions):
-            """computes the faithfulness metric"""
+            """Computes the faithfulness metric."""
             X2 = self.interpret_computation_steps(wavs)[0]
 
             _, predictions_masked, _ = self.classifier_forward(X2)
@@ -465,10 +501,10 @@ class InterpreterESC50Brain(sb.core.Brain):
             predictions = F.softmax(predictions, dim=1)
             predictions_masked = F.softmax(predictions_masked, dim=1)
 
-            # get the prediction indices
+            # Get the prediction indices
             pred_cl = predictions.argmax(dim=1, keepdim=True)
 
-            # get the corresponding output probabilities
+            # Get the corresponding output probabilities
             predictions_selected = torch.gather(
                 predictions, dim=1, index=pred_cl
             )
@@ -492,7 +528,7 @@ class InterpreterESC50Brain(sb.core.Brain):
     def on_stage_end(self, stage, stage_loss, epoch=None):
         """Gets called at the end of an epoch.
         Plots in subplots the values of `self.batch_to_plot` and saves the
-        plot to the experiment folder. `self.hparams.output_folder`"""
+        plot to the experiment folder `self.hparams.output_folder`."""
         current_fid = self.top_3_fidelity.summarize("average")
         test_stats = {
             "acc": self.acc_metric.summarize("average"),
@@ -503,15 +539,14 @@ class InterpreterESC50Brain(sb.core.Brain):
             "faithfulness_mean": torch.Tensor(self.faithfulness.scores).mean(),
         }
 
-        # The train_logger writes a summary to stdout and to the logfile.
+        # The train_logger writes a summary to stdout and to the log file
         self.hparams.train_logger.log_stats(
             stats_meta={"epoch": epoch}, test_stats=test_stats
         )
 
 
 def dataio_prep(hparams):
-    "Creates the datasets and their data processing pipelines."
-
+    """Creates the datasets and their data processing pipelines."""
     data_audio_folder = hparams["audio_data_folder"]
     config_sample_rate = hparams["sample_rate"]
     label_encoder = sb.dataio.encoder.CategoricalEncoder()
@@ -553,7 +588,7 @@ def dataio_prep(hparams):
     @sb.utils.data_pipeline.takes("class_string")
     @sb.utils.data_pipeline.provides("class_string", "class_string_encoded")
     def label_pipeline(class_string):
-        """the label pipeline"""
+        """The label pipeline."""
         yield class_string
         class_string_encoded = label_encoder.encode_label_torch(class_string)
         yield class_string_encoded
@@ -588,6 +623,9 @@ def dataio_prep(hparams):
 
 
 if __name__ == "__main__":
+    # This flag enables the built-in cuDNN auto-tuner
+    # torch.backends.cudnn.benchmark = True
+
     # CLI:
     hparams_file, run_opts, overrides = sb.parse_arguments(sys.argv[1:])
